// src/lib/api/generated/model/profileMeResDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * BE SM Service API
 * BE SM Service API
 * OpenAPI spec version: 1.0.0
 */
import type { ProfileMeResDtoExternalIds } from './profileMeResDtoExternalIds';
import type { ProfileMeResDtoRole } from './profileMeResDtoRole';
import type { ProfileMeResDtoSocialNetworks } from './profileMeResDtoSocialNetworks';
import type { ProvinceResDto } from './provinceResDto';
import type { ShopResDto } from './shopResDto';
import type { UserBankShortenedResDto } from './userBankShortenedResDto';
import type { WardResDto } from './wardResDto';

export interface ProfileMeResDto {
  id: string;
  accountType?: string;
  avatarUrl?: string;
  idCardNumber?: string;
  issuingAuthority?: string;
  dateOfIssue?: string;
  fullName: string;
  phoneNumber?: string;
  email?: string;
  taxId?: string;
  representative?: string;
  representativeRole?: string;
  businessLicenseId?: string;
  userBank?: UserBankShortenedResDto;
  externalIds?: ProfileMeResDtoExternalIds;
  provinceId?: string;
  province?: ProvinceResDto;
  wardId?: string;
  ward?: WardResDto;
  address?: string;
  socialNetworks?: ProfileMeResDtoSocialNetworks;
  isNeedUpdateProfile?: boolean;
  isNeedChangePassword?: boolean;
  referralUrl?: string;
  referralCode?: string;
  role?: ProfileMeResDtoRole;
  isEKYCVerified?: boolean;
  shop?: ShopResDto;
}