// src/lib/api/generated/model/userRefResDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * BE SM Service API
 * BE SM Service API
 * OpenAPI spec version: 1.0.0
 */
import type { UserProfileRefResDto } from './userProfileRefResDto';
import type { UserRefResDtoRole } from './userRefResDtoRole';
import type { UserRefResDtoStatus } from './userRefResDtoStatus';

export interface UserRefResDto {
  id: string;
  username?: string;
  role: UserRefResDtoRole;
  status: UserRefResDtoStatus;
  profile?: UserProfileRefResDto;
  code?: string;
}